+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for FastQC (#5)
 - Support for MEGAHIT (#13)
 - Support for Unicycler (#14)
+- Support for assembly configuration file (#16)
 
 
 ## [0.1] 2021-12-01